backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-multipart==0.0.19
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pydantic-settings==2.6.1
